requirements.txt
tensorflow>=2.15.0
streamlit>=1.31.0
pandas>=2.2.0
numpy>=1.24.0
plotly>=5.18.0
python-binance>=1.0.19
scikit-learn>=1.4.0 